requirements-dev.txt
azure-functions
pytest
pyhamcrest
